MusicApp/requirements.txt
click==6.6
colorama==0.3.3
configparser==3.5.0
coverage==4.0a5
cssselect==0.9.2
decorator==3.4.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.1
django-bootstrap3==7.1.0
django-filter==1.0.1
django-localflavor==1.4.1
django-tastypie==0.13.3
djangorestframework==3.5.3
Pillow==4.0.0
requests==2.11.1
requests-futures==0.9.7
SQLAlchemy==0.8.4
virtualenv==12.0.7
w3lib==1.15.0
zope.interface==4.3.2
